requirements.txt
# Phase 2 RAG - pinned versions for reproducibility
# PDF
PyMuPDF==1.24.10

# Embeddings & vector index (local, no API)
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1

# Hybrid retrieval: BM25
rank-bm25==0.2.2

# Utilities (protobuf needed for transformers/TF compatibility in some envs)
numpy>=1.24.0,<3.0.0
tqdm>=4.66.0
protobuf>=4.21.0,<6

# Optional: load .env for API keys
python-dotenv>=1.0.0

# Optional: for manifest/eval CSV
# (stdlib csv, json sufficient; pandas optional)
pandas>=2.0.0

# Testing
pytest>=7.0.0
